Dev kwic

+++ src/quanteda.h
@@ -12,9 +12,9 @@ using namespace tbb;
 namespace quanteda{
 
     inline String join(CharacterVector &tokens, String &delim){
-        if(tokens.size() == 0) return "";
+        if (tokens.size() == 0) return "";
         String token = tokens[0];
-        for (int i = 1; i < tokens.size(); i++) {
+        for (size_t i = 1; i < tokens.size(); i++) {
           token += delim;
           token += tokens[i];
           //Rcout << "Joined " << token.get_cstring()  << "\n";
@@ -24,28 +24,28 @@ namespace quanteda{
       }
     
     inline std::string join(std::vector< std::string > &tokens, std::string &delim){
-        if(tokens.size() == 0) return "";
+        if (tokens.size() == 0) return "";
         std::string token = tokens[0];
-        for (int i = 1; i < tokens.size(); i++) {
+        for (size_t i = 1; i < tokens.size(); i++) {
           token += delim + tokens[i];
         }
         return token;
     }
 
     inline bool has_na(IntegerVector vec) {
-        for (int i = 0; i < vec.size(); ++i) {
-            if(vec[i] == NA_INTEGER) return true;
+        for (size_t i = 0; i < vec.size(); ++i) {
+            if (vec[i] == NA_INTEGER) return true;
         }
        return false;
     }
 
     inline List as_list(std::vector< std::vector<unsigned int> > &tokens){
         List list(tokens.size());
-        for(int h = 0; h < tokens.size(); h++){
-            if(tokens[h].size() > 0){
+        for (size_t h = 0; h < tokens.size(); h++) {
+            if (tokens[h].size() > 0) {
                 IntegerVector temp = Rcpp::wrap(tokens[h]);
                 list[h] = temp;
-            }else{
+            } else {
                 list[h] = R_NilValue;
             }
             
@@ -59,15 +59,13 @@ namespace ngrams {
     typedef std::vector<unsigned int> Ngram;
     typedef std::vector< std::vector<unsigned int> > Ngrams;
     typedef std::vector<unsigned int> Text;
-    //typedef std::vector<int> Text;
     typedef std::vector< std::vector<unsigned int> > Texts;
-    //typedef std::vector< std::vector<int> > Texts;
     struct hash_ngram {
         std::size_t operator() (const Ngram &vec) const {
             //unsigned int seed = std::accumulate(vec.begin(), vec.end(), 0);
             //return std::hash<unsigned int>()(seed);
             unsigned int seed = 0;
-            for(int i = 0; i < vec.size(); i++){
+            for (size_t i = 0; i < vec.size(); i++) {
                 seed += vec[i] << (i * 8); // shift elements 8 bit
             }
             return std::hash<unsigned int>()(seed);
